Replace reset behavior when re-applying the patch routine of a Rack.
The replacement is tracking of inputs that have been set within the
patch routine. The execution of the patch routine is bookended with
method calls to each module in the Rack. Inputs that have been exposed
on a module, but haven't been set during the patch phase will be set to
their normalized value.